+++ src/Doctrine/Common/State/PersistProcessor.php
@@ -50,17 +50,35 @@ final class PersistProcessor implements ProcessorInterface
         // https://github.com/doctrine/orm/issues/8461#issuecomment-1250233555
         if ($operation instanceof HttpOperation && 'PUT' === $operation->getMethod() && ($operation->getExtraProperties()['standard_put'] ?? false)) {
             \assert(method_exists($manager, 'getReference'));
-            // TODO: the call to getReference is most likely to fail with complex identifiers
             $newData = $data;
-            if ($previousData = $context['previous_data']) {
-                $newData = 1 === \count($uriVariables) ? $manager->getReference($class, current($uriVariables)) : clone $previousData;
-            }
-
             $identifiers = array_reverse($uriVariables);
             $links = $this->getLinks($class, $operation, $context);
             $reflectionProperties = $this->getReflectionProperties($data);
 
-            if (!$previousData) {
+            // TODO: the call to getReference is most likely to fail with complex identifiers
+            if ($previousData = $context['previous_data']) {
+                $classMetadata = $manager->getClassMetadata($class);
+                $identifiers = $classMetadata->getIdentifierValues($previousData);
+                $newData = 1 === \count($identifiers) ? $manager->getReference($class, current($identifiers)) : clone $previousData;
+
+                foreach ($reflectionProperties as $propertyName => $reflectionProperty) {
+                    // // Don't override the property if it's part of the subresource system
+                    if (isset($identifiers[$propertyName]) || isset($uriVariables[$propertyName])) {
+                        continue;
+                    }
+
+                    // Skip URI variables as sometime an uri variable is not the doctrine identifier
+                    foreach ($links as $link) {
+                        if (\in_array($propertyName, $link->getIdentifiers(), true)) {
+                            continue 2;
+                        }
+                    }
+
+                    if (($newValue = $reflectionProperty->getValue($data)) !== $reflectionProperty->getValue($newData)) {
+                        $reflectionProperty->setValue($newData, $newValue);
+                    }
+                }
+            } else {
                 foreach (array_reverse($links) as $link) {
                     if ($link->getExpandedValue() || !$link->getFromClass()) {
                         continue;
@@ -72,24 +90,6 @@ final class PersistProcessor implements ProcessorInterface
                     foreach ($identifierProperties as $identifierProperty) {
                         $reflectionProperty = $reflectionProperties[$identifierProperty];
                         $reflectionProperty->setValue($newData, $this->getIdentifierValue($identifiers, $hasCompositeIdentifiers ? $identifierProperty : null));
-                    }
-                }
-            } else {
-                foreach ($reflectionProperties as $propertyName => $reflectionProperty) {
-                    // Don't override the property if it's part of the subresource system
-                    if (isset($uriVariables[$propertyName])) {
-                        continue;
-                    }
-
-                    foreach ($links as $link) {
-                        $identifierProperties = $link->getIdentifiers();
-                        if (\in_array($propertyName, $identifierProperties, true)) {
-                            continue;
-                        }
-
-                        if (($newValue = $reflectionProperty->getValue($data)) !== $reflectionProperty->getValue($newData)) {
-                            $reflectionProperty->setValue($newData, $newValue);
-                        }
                     }
                 }
             }
@@ -113,7 +113,7 @@ final class PersistProcessor implements ProcessorInterface
     private function isDeferredExplicit(DoctrineObjectManager $manager, $data): bool
     {
         $classMetadata = $manager->getClassMetadata($this->getObjectClass($data));
-        if (method_exists($classMetadata, 'isChangeTrackingDeferredExplicit')) {
+        if ($classMetadata && method_exists($classMetadata, 'isChangeTrackingDeferredExplicit')) { // @phpstan-ignore-line metadata can be null
             return $classMetadata->isChangeTrackingDeferredExplicit();
         }